==> proteomics/10_SWATH_preprocessing.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd
import numpy as np
from scipy import optimize
from scipy.stats import sem

import os
import sys

import matplotlib.pyplot as plt
import matplotlib.ticker
from venn import venn, pseudovenn
import seaborn as sns
from matplotlib.collections import PathCollection

from Bio import SeqIO, SeqUtils, Seq
from Bio.SeqUtils.ProtParam import ProteinAnalysis

from addict import Dict
import json

sys.path.append("../")
import plotting as my_plot

sys.path.append("./mean_field_elect/")
import proteomics_functions as prot_fun
import base_classes
import morbidelli as m

# Load data

dfs = Dict()
dfs_mab = Dict()

file = pd.ExcelFile(f'./data/Chase SWATH 20220830.xlsx')
dfs.hccf.feed  = file.parse('HCCF-Feed')
dfs.hccf.large = file.parse('HCCF-LrgAgg')
dfs.hccf.small = file.parse('HCCF-SmAgg')
dfs.hccf.mab   = file.parse('HCCF-mAb')
dfs.hccf.spf1  = file.parse('HCCF-LMW1')
dfs.hccf.spf2  = file.parse('HCCF-LMW2') 
dfs.pavin.feed  = file.parse('PAFVIN-Feed')
dfs.pavin.large = file.parse('PAFVIN-LrgAgg')
dfs.pavin.small = file.parse('PAFVIN-SmAgg')

file = pd.ExcelFile('./data/Chase SWATH 20221128.xlsx')
dfs.pavin.mab = file.parse('PAFVIN mAb!')


for feed in ['hccf', 'pavin']:
    for frac in dfs[feed].keys():
        dfs[feed][frac].columns = ['accn', 'name', 'rep1_log2_norm_area', 'rep2_log2_norm_area', 'rep3_log2_norm_area', 'prot_mw', 'rep1_ng', 'rep2_ng', 'rep3_ng', 'ave_ng', 'cv']
        if frac == 'spf2':
            dfs[feed][frac]['ave_ppm'] = dfs[feed][frac]['ave_ng']/5.0 * 1e3 # x ng/5 ug * 1e-3 ug/ng * 1e6 ppm
        else:
            dfs[feed][frac]['ave_ppm'] = dfs[feed][frac]['ave_ng']/90.91 * 1e3 # x ng/90.91 ug * 1e-3 ug/ng * 1e6 ppm
                    
        # Get mAb as a separate df
        dfs_mab[feed][frac] = dfs[feed][frac][dfs[feed][frac]['name'].str.contains('Custom')]
        dfs_mab[feed][frac].reset_index(inplace=True, drop=True)
        
        # Select only CHO HCPs
        dfs[feed][frac] = dfs[feed][frac][dfs[feed][frac]['name'].str.contains('Cricetulus griseus')]
        dfs[feed][frac].reset_index(inplace=True, drop=True)

# Assemble a flat df with ave_ppm values

names = {}

for source in dfs.keys():
    for frac in dfs[source].keys():
        df = dfs[source][frac]
        for i, cont in df.iterrows():
            names[cont['accn']] = cont['name']
                
df_master = pd.DataFrame.from_dict(names, orient='index')
df_master.reset_index(inplace=True)
df_master.columns = ['accn', 'name']

for source in dfs.keys():
    for frac in dfs[source].keys():
        df = dfs[source][frac]
        for i, cont in df.iterrows():
            df_master.loc[df_master.accn == cont.accn, f'{source}_{frac}'] = cont.ave_ppm

df_master.to_csv('./generated_tables/swath_master_df_ppm_with_na.csv', index=False)

df_master = df_master.fillna(0)

df_master.to_csv('./generated_tables/swath_master_df_ppm.csv', index=False)

df_master.head()

# Add biophysical properties

# # Get new html links to look up sequences

# html = ''
# cnt = 0

# for i, a in enumerate(df_master.accn):
#     if i % 200 == 0:
#         print(html[:-1], '\n'*2)
#         html = 'https://www.ncbi.nlm.nih.gov/protein/'        
#     html += a + ','
#     cnt += 1
    
# print(html[:-1], '\n'*2)

# My substitution rules for uncertain amino acids
my_sub_rules = {'B':'D', 'Z':'E', 'X':'A', 'J':'L'}

# Get sequence dictionary {accession:sequence_object}
sequences = {}
subbed_ids = []

for r in SeqIO.parse("./data/sequences_swath.fasta", "fasta"):
    for aa in my_sub_rules.keys(): # uncertain amino acids
        if aa in r.seq:
            r.seq = Seq.Seq(str(r.seq).replace(aa, my_sub_rules[aa]))
            subbed_ids.append(r.id)
    sequences[r.id] = r.seq

missing = []
for i, cont in df_master.iterrows():
    if cont.accn not in list(sequences.keys()):
        missing.append(cont.accn)
assert len(missing) == 0

# Get pI and mass dictionaries {accession:pI/mass}
pI_vals = {}
masses = {}

for p_id, seq in sequences.items():
    pI, is_solved = prot_fun.get_pI(seq)
    assert is_solved
    pI_vals[p_id] = pI
    masses[p_id] = SeqUtils.molecular_weight(seq, seq_type='protein')

# Add sequences, pI values, and masses to df_master
for i, cont in df_master.iterrows():
    df_master.at[i, 'sequence'] = str(sequences[cont.accn])
    df_master.at[i, 'pI'] = pI_vals[cont.accn]
    df_master.at[i, 'mass'] = masses[cont.accn]

# Get other biophysical property dictionaries (assuming pH 7.0)
net_charges, net_neg_charges, net_pos_charges, charge_densities, charge_densities_neg, charge_densities_pos = {}, {}, {}, {}, {}, {}

for p_id, seq in sequences.items():
    net_charge, net_neg_charge, net_pos_charge, charge_dens, charge_dens_neg, charge_dens_pos = prot_fun.get_charge(pH=7.0, seq=seq, charge_contributions=True)
    net_charges[p_id] = net_charge
    net_neg_charges[p_id] = net_neg_charge
    net_pos_charges[p_id] = net_pos_charge
    charge_densities[p_id] = charge_dens
    charge_densities_neg[p_id] = charge_dens_neg
    charge_densities_pos[p_id] = charge_dens_pos

# Add these biophysical properties to df_master
for i, cont in df_master.iterrows():
    df_master.at[i, 'net_charge'] = net_charges[cont.accn]
    df_master.at[i, 'net_charge_neg'] = net_neg_charges[cont.accn]
    df_master.at[i, 'net_charge_pos'] = net_pos_charges[cont.accn]
    df_master.at[i, 'charge_dens_C_m2'] = charge_densities[cont.accn]
    df_master.at[i, 'charge_dens_neg_C_m2'] = charge_densities_neg[cont.accn]
    df_master.at[i, 'charge_dens_pos_C_m2'] = charge_densities_pos[cont.accn]

df_master.to_csv('./generated_tables/swath_master_df_ppm_with_properties.csv', index=False)
